% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipesheet.R
\name{edit}
\alias{edit}
\title{edit - Function to Edit Excel Files}
\usage{
edit(data)
}
\arguments{
\item{data}{A data frame to be written to Excel.}
}
\value{
A tibble containing the edited data.
}
\description{
This function writes the input data frame to an Excel file, opens it for editing in either LibreOffice (for Unix-based systems) or Microsoft Excel (for Windows systems), and returns the edited data as a tibble.
}
\examples{
data(mtcars)
edited_data <- edit(mtcars)
}
