% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipesheet.R
\name{view}
\alias{view}
\title{view - Function to View Excel Files}
\usage{
view(data, row.names = F)
}
\arguments{
\item{data}{A data frame to be written to Excel.}

\item{row.names}{Logical value indicating whether row names should be included in the Excel file. Default is \code{FALSE}.}
}
\description{
This function writes the input data frame to an Excel file and opens it in either LibreOffice (for Unix-based systems) or Microsoft Excel (for Windows systems) for viewing.
}
\examples{
data(mtcars)
view(mtcars)
}
